+++ patterns/sine.py
@@ -16,9 +16,9 @@ class Pattern(object):
         self.cube.clear()
 # sine needs 0 to be midpoint... ah. but there's no midpoint
         for x in range(0, self.cube.size):
-            y = math.sin(x+self.counter) * self.cube.size/2;
+            y = int(math.sin(x+self.counter) * self.cube.size/2);
             y += self.cube.size/2
-            z = math.sin(x+self.counter-1) * self.cube.size/2;
+            z = int(math.sin(x+self.counter-1) * self.cube.size/2);
             z += self.cube.size/2
             color = cubehelper.random_color()
             self.cube.set_pixel([x,z,y],color);
